+++ src/channels/alipay_qr.js
@@ -4,8 +4,8 @@ var callbacks = require('../callbacks');
 /*global AlipayJSBridge*/
 module.exports = {
 
-  handleCharge: function (charge) {
-    var credential = charge.credential[charge.channel];
+  handleCharge: function (payment) {
+    var credential = payment.credential[payment.channel];
     if (hasOwn.call(credential, 'transaction_no')) {
       this.tradePay(credential.transaction_no);
     } else {