+++ src/channels/bfb_wap.js
@@ -4,9 +4,9 @@ var hasOwn = {}.hasOwnProperty;
 
 module.exports = {
 
-  handleCharge: function(charge) {
-    var channel = charge.channel;
-    var credential = charge.credential[channel];
+  handleCharge: function(payment) {
+    var channel = payment.channel;
+    var credential = payment.credential[channel];
 
     if (!hasOwn.call(credential, 'url')) {
       callbacks.innerCallback('fail',