+++ src/channels/cmb_pc_qr.js
@@ -3,9 +3,9 @@ var hasOwn = {}.hasOwnProperty;
 
 module.exports = {
 
-  handleCharge: function(charge) {
-    var channel = charge.channel;
-    var credential = charge.credential[channel];
+  handleCharge: function(payment) {
+    var channel = payment.channel;
+    var credential = payment.credential[channel];
     var baseURL;
     if (hasOwn.call(credential, 'channel_url')) {
       baseURL = credential.channel_url;