+++ src/channels/cmb_wallet.js
@@ -6,8 +6,8 @@ module.exports = {
   CMB_WALLET_URL:
     'https://netpay.cmbchina.com/netpayment/BaseHttp.dll?MB_EUserPay',
 
-  handleCharge: function(charge) {
-    var credential = charge.credential[charge.channel];
+  handleCharge: function(payment) {
+    var credential = payment.credential[payment.channel];
     var request_url = this.CMB_WALLET_URL;
     if (hasOwn.call(credential, 'ChannelUrl')) {
       request_url = credential.ChannelUrl;