+++ src/channels/jdpay_wap.js
@@ -7,8 +7,8 @@ module.exports = {
   JDPAY_H5_URL: 'https://h5pay.jd.com/jdpay/saveOrder',
   JDPAY_PC_URL: 'https://wepay.jd.com/jdpay/saveOrder',
 
-  handleCharge: function(charge) {
-    var credential = charge.credential[charge.channel];
+  handleCharge: function(payment) {
+    var credential = payment.credential[payment.channel];
     var request_url = this.JDPAY_H5_URL;
     if (hasOwn.call(credential, 'channelUrl')) {
       request_url = credential.channelUrl;