+++ src/channels/paypal.js
@@ -1,8 +1,8 @@
 var utils = require('../utils');
 
 module.exports = {
-  handleCharge: function(charge) {
-    var credential = charge.credential[charge.channel];
-    utils.redirectTo(credential, charge.channel);
+  handleCharge: function(payment) {
+    var credential = payment.credential[payment.channel];
+    utils.redirectTo(credential, payment.channel);
   }
 };