+++ src/channels/qpay_pub.js
@@ -8,8 +8,8 @@ module.exports = {
   SRC_URL: 'https://open.mobile.qq.com/sdk/qqapi.js?_bid=152',
   ID: 'mqq_api',
 
-  handleCharge: function (charge) {
-    var credential = charge.credential[charge.channel];
+  handleCharge: function (payment) {
+    var credential = payment.credential[payment.channel];
 
     if (!hasOwn.call(credential, 'token_id')) {
       callbacks.innerCallback('fail',