+++ src/channels/upacp_pc.js
@@ -4,8 +4,8 @@ module.exports = {
 
   UPACP_PC_URL: 'https://gateway.95516.com/gateway/api/frontTransReq.do',
 
-  handleCharge: function(charge) {
-    var credential = charge.credential[charge.channel];
+  handleCharge: function(payment) {
+    var credential = payment.credential[payment.channel];
     utils.formSubmit(this.UPACP_PC_URL, 'post', credential);
   }
 };