+++ src/channels/wx_wap.js
@@ -4,14 +4,14 @@ var hasOwn = {}.hasOwnProperty;
 
 module.exports = {
 
-  handleCharge: function(charge) {
-    var credential = charge.credential[charge.channel];
+  handleCharge: function(payment) {
+    var credential = payment.credential[payment.channel];
     if (typeof credential === 'string') {
-      utils.redirectTo(credential, charge.channel);
+      utils.redirectTo(credential, payment.channel);
     } else if (typeof credential === 'object'
       && hasOwn.call(credential, 'url')
     ) {
-      utils.redirectTo(credential.url, charge.channel);
+      utils.redirectTo(credential.url, payment.channel);
     } else {
       callbacks.innerCallback('fail',
         callbacks.error('invalid_credential', 'credential 格式不正确'));