+++ src/payment_elements.js
@@ -17,51 +17,51 @@ module.exports = {
   time_expire: null,
 
   init: function (charge_or_order) {
-    var charge;
+    var payment;
     if (typeof charge_or_order === 'string') {
       try {
-        charge = JSON.parse(charge_or_order);
+        payment = JSON.parse(charge_or_order);
       } catch (err) {
         callbacks.innerCallback('fail',
           callbacks.error('json_decode_fail', err));
         return;
       }
     } else {
-      charge = charge_or_order;
+      payment = charge_or_order;
     }
 
-    if (typeof charge === 'undefined') {
+    if (typeof payment === 'undefined') {
       callbacks.innerCallback('fail', callbacks.error('json_decode_fail'));
       return;
     }
 
-    if (hasOwn.call(charge, 'object') && charge.object == 'order') {
-      charge.or_id = charge.id;
-      charge.order_no = charge.merchant_order_no;
-      var charge_essentials = charge.charge_essentials;
-      charge.channel = charge_essentials.channel;
-      charge.credential = charge_essentials.credential;
-      charge.extra = charge_essentials.extra;
-      if(hasOwn.call(charge, 'charge') && charge.charge != null) {
-        charge.id = charge.charge;
+    if (hasOwn.call(payment, 'object') && payment.object == 'order') {
+      payment.or_id = payment.id;
+      payment.order_no = payment.merchant_order_no;
+      var charge_essentials = payment.charge_essentials;
+      payment.channel = charge_essentials.channel;
+      payment.credential = charge_essentials.credential;
+      payment.extra = charge_essentials.extra;
+      if(hasOwn.call(payment, 'payment') && payment.payment != null) {
+        payment.id = payment.payment;
       } else if(hasOwn.call(charge_essentials, 'id')
         && charge_essentials.id != null) {
-        charge.id = charge_essentials.id;
-      } else if(hasOwn.call(charge, 'charges')) {
-        for(var i = 0; i < charge.charges.data.length; i++){
-          if(charge.charges.data[i].channel === charge_essentials.channel) {
-            charge.id = charge.charges.data[i].id;
+        payment.id = charge_essentials.id;
+      } else if(hasOwn.call(payment, 'charges')) {
+        for(var i = 0; i < payment.charges.data.length; i++){
+          if(payment.charges.data[i].channel === charge_essentials.channel) {
+            payment.id = payment.charges.data[i].id;
             break;
           }
         }
       }
-    } else if(hasOwn.call(charge, 'object') && charge.object == 'recharge') {
-      charge = charge.charge;
+    } else if(hasOwn.call(payment, 'object') && payment.object == 'recharge') {
+      payment = payment.payment;
     }
 
     for (var key in this) {
-      if (hasOwn.call(charge, key)) {
-        this[key] = charge[key];
+      if (hasOwn.call(payment, key)) {
+        this[key] = payment[key];
       }
     }
     return this;