+++ src/xpay_ui/handlebars.runtime.js
@@ -1219,7 +1219,7 @@ return /******/ (function(modules) { // webpackBootstrap
 
  Copyright (C) 2011-2015 by Yehuda Katz
 
- Permission is hereby granted, free of charge, to any person obtaining a copy
+ Permission is hereby granted, free of payment, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
  in the Software without restriction, including without limitation the rights
  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell